[front] Migrate missing_action_catcher to new MCP server structure
Move to lib/api/actions/servers/ with metadata.ts separation to avoid
pulling backend dependencies (logger, statsDClient) into frontend code
via constants.ts imports.
